update file structure + commit to dynamic routes
