+++ src/app/(explore)/[category]/page.tsx
@@ -2,9 +2,9 @@ import { Suspense } from 'react';
 import { notFound } from 'next/navigation';
 
 import { PreloadQuery } from '@/lib/graphql/apollo-client';
-import { getCategoryConfig, getDynamicRoutes } from '@/config/category';
-import ListLayout from '@/components/layouts/list-layout';
+import { getCategoryConfig, getCategoryRoutes } from '@/config/category';
 import DynamicList from './_components/dynamic-list';
+import Search from '@/components/ui/search';
 
 type Props = {
   params: { 
@@ -15,7 +15,7 @@ type Props = {
 // export const dynamicParams = false;
 
 export const generateStaticParams = () => {
-  const routes = getDynamicRoutes();
+  const routes = getCategoryRoutes();
   return routes.map((path) => ({
     category: path 
   }));
@@ -35,12 +35,21 @@ export default async function Page({ params: { category } }: Props) {
   }
 
   return (
-    <ListLayout>
+    <section className="m-10">
+      <header className="border-b-2 font-semibold">
+        <p>Explore {category}</p>
+        <br/>
+        <Suspense>
+          <Search />
+        </Suspense>
+      </header>
+      <div>
       <PreloadQuery query={QUERY_CATEGORY_ITEMS}>
         <Suspense fallback={<p>LOADING...</p>}>
           <DynamicList />
         </Suspense>
       </PreloadQuery>
-    </ListLayout>
+      </div>
+    </section>
   );
 }
