+++ src/components/equipment/index.ts
@@ -1,0 +1,2 @@
+export * from './equipment-card';
+export * from './equipment-listing';
