+++ src/components/magic-items/index.ts
@@ -1,0 +1,2 @@
+export * from './magic-item-card';
+export * from './magic-item-listing';
