+++ src/components/monsters/index.ts
@@ -1,0 +1,2 @@
+export * from './monster-card';
+export * from './monster-listing';
