+++ src/components/spells/index.ts
@@ -1,0 +1,2 @@
+export * from './spell-card';
+export * from './spell-listing';
